refactor: postgres result parser file.
